% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Generate a Correlation Matrix}
\usage{
correlation_matrix(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A correlation matrix.
}
\description{
This function generates a correlation matrix for all numeric variables in a dataset.
}
\examples{
data(mtcars)
correlation_matrix(mtcars)
}
