% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot Distribution of Variables}
\usage{
plot_distribution(data, column)
}
\arguments{
\item{data}{A data frame.}

\item{column}{The name of the column to plot.}
}
\value{
A ggplot object.
}
\description{
This function plots the distribution of numeric variables in a dataset.
}
\examples{
data(mtcars)
plot_distribution(mtcars, "mpg")
}
